More details to lower the threshold for beginners
Some parts have been changed or explained more explicitly.
The intention was to lower the threshold for beginners.
This means, some more explanation of the namespace usage as not every
developer is currently 100% familiar with its usage and we should rather
take them by the hand to enjoy namespaces than scare them off.