[book][page_creation] Updating the page creation chapter to be consistent with the Standard Distribution
This is primarily in response to the Sensio namespace being used in the Standard Distribution and being mapped to a directory in the autoloader. In order to avoid autoloading problems, it's easier to use the "Acme" namespace in our examples and map it to the ``src`` directory. The Standard Distribution does this already, so this matches it.

Additionally, the chapter needed more explanation up front about how to actually get started (in case someone is actually coding along with the tutorial). So, a section on creating the bundle was added early in the chapter. Overall, we'll see how it reads.