[reference] Rewrite assets_version_format docs and add tip for path prefixing
Versioned path prefixes are useful for CDN's that do not support cache-busting via query strings. This seemed small enough to add as a tip, although it might be better expanded into a cookbook article (and this tip reduced to a "see cookbook" link).

Also added a note about required escaping of percentage signs in the option value.