[#1810] Adding a Learn More section to the console introduction
This introduces the idea of having short "intro" chapters for each component, followed by lots of related docs. This is a model that we largely want to follow in the components and in the framework book itself (we also need more in-context cross-linking between documents).

I've also updated a few document titles to be more descriptive.