Removing several "2.1" references, which is consistent with our release process:
"Remove all versionadded directives - and any other notes related to features changing or being new - for the version (e.g. 2.1) from the master branch. The result is that the next release (which is the first that comes entirely after the end of maintenance of this branch), will have no mentions of the old version (e.g. 2.1)."

This should have been done when 2.1 reached end of maintenance, but the release process wasn't formed at that time.

Also, the release of 2.3 and the end of maintenance for 2.1 were unique because they basically happened at the same time. For our purposes, we assume that 2.3 was released before the 2.1 end of life, which means that the 2.1 references are removed for 2.4 (master), not for 2.3.