make the advantage and disadvantage of the ACL more clear
For me as a SF2 developer to handle data manipulation was pretty new to me. So I asked in the irc and searched the web and finally landed on this documentation page. So it looked like this is the way to go and it is a usual thing like using bundles in SF2. But it isnt, it is a far more complex permission handler which will work upfront and is not that much flexible.

I would highly recommend to add at least some notes, so that other developers getting a note on that and can maybe save some days of work as the opposite of me.