Removed all occurences of $this->getRequest()
The request is a data object, and so the request should not be
inside the container. That's why the request will be removed
from the container in Symfony 3.0

So it is better to let the HttpKernel injects the request when
the user want to use the request.

More over, it is faster. The HttpKernel will always check if the
request should be injected. So injecting the request is time constant.
However, retrieve the request from the container will execute more
calls.

Conflicts:
	book/controller.rst
	cookbook/doctrine/file_uploads.rst
	cookbook/doctrine/registration_form.rst
	cookbook/form/form_collections.rst
	quick_tour/the_controller.rst