Removed duplicated DQL documentation
The doctrine docs already has a big great documentation on this. There is
no reason to duplicate it in the sf docs. This commit removes everything
except a simple example.

It also moves the sentence about the return result above, since that makes
more sense now the getSingleResult example is removed.