[#3596] Tweaking language on HTTP cache
I realized that the sentence on the notes was a bid hard to read. I also removed a usage of
the word "client" right before this, so I can use the word "cache" instead (meaning, I want
the user to be thinking about the reverse proxy, not a user's browser cache - that's valid,
but not the typical use-case).