[book] [routing] fixed a note that wasn't properly updated
Previously, the code used the generate() method of the service and
the note explained the generateUrl() method of the base controller.
Then, the example was updated to use the generateUrl(), but the
note kept explaining the generateUrl(). This change is just about
explaining in the note the alterantive generate() method of the service.