Reorder page from simple to advanced
As a followup on https://github.com/symfony/symfony/pull/10980 I want to propose to reorder the page as the obvious way should come first. I took me way to long to find the correct way :p

I added two headers

- The default ExceptionController
- Override error templates

in order to introduce how override works and moved the controller override to the bottom as that is more advanced.