Removed the redundant usage of layer.
Since DBAL is database abstraction layer, and is defined in the document as such, to me it reads funny to have it as database abstraction layer layer. It sounds nitpicky, I know, but I find it an easier read this way. Also, the possessive DBAL stuck out ("DBAL's layer") as the DBAL doesn't own a layer, it is the layer.