Fix description of ConfirmationQuestion
In my testing and in my brief reading of the Symfony\Component\Console\Question code, the old description of re-asking the question does not occur and I see no way for that to happen.  That would be nice, desirable behavior.  However, the truth is the ask() method always returns ``true`` or ``false`` upon any input.