[RFC] Clarification on formatting for bangs (!)
It seems to me that 
> Add a single space around operators

would mean using the practice on the bang operator as well; as discussed in [this thread](https://github.com/php-fig/fig-standards/issues/102), there is somewhat of a conflict--so this may not be the way to go. If the PSR could be amended, then this change would make sense.