Update page_creation.rst
Probably useful to instruct folks to start their new file with <?php --maybe not always, but in this instance The Book goes straight from telling the person to create a new file to giving the exact contents of the file. In truth, I overlooked the error, as did the kind person helping me on irc #symfony, for a good 30 min.