Update page_creation.rst
Radical proposal #1: focus more specifically on creating pages in this document and put related (but not precisely 'dependent') discussions of directories, bundles and environments in a separate document called Structuring a Symfony App. 

Radical proposal #2: take a heavy axe to The Book to eliminate every description, explanation and detail that's not at the highest possible altitude. Replace extraneous detail with links to the deeper information elsewhere, e.g. Best Practices. The ideal outcome would be DRYer docs and an easier to read and maintain The Book.

Keep in mind that I'm a person with 'fresh eyes' vs one with a clear understanding of all objectives that must be met on this introductory page. I expect the best course is something between what I propose and what is currently there.

That said, this revision takes an initial stab, so please feel free to use or ignore as you see fit.