Update the_big_picture.rst
When following the steps in this documentation, I ended up with this: http://i.imgur.com/nK3XM2B.png
Got the suggestion to install the assets on the IRC channel and it did the trick for me. This addition to the docs might help others facing the same problem to make their app look as expected.