Removing a section about Roles that I think has no real use-case
It was added 4 years ago at sha: 5a2824b6, and it's original purpose was to talk
about Roles as domain objects. If this has a real-use case, then we should clarify
what it is, and move it to its own cookbook. This is a details that beginners are
trying to deal with, unnecessarily.