Add note about using choice_label as a callback
I was wondering how to call a method with an argument for some time and then I've came with an idea of using ArrayAccess that can be used with PropertyAccessor.

Then I found I can actually use callback in `choice_label`, but this was not documented. Is this correct usage?