Added hints to spool config section
The former explanation suggests that a single file is used for spooling, which is not the case of course. But I ran into that problem and since Symfony falls back to memory if the dir is not writable (as it seems) with no error message, this addition might help some people to get spools running.