reflect behavior changes in cache generation
Starting with Symfony 2.8, the class cache is generated when the cache
is warmed up. This means that you cannot prevent the cache from being
generated, but you must simply not load it to ease debugging. Also, this
means that you have to ignore the cache files in your IDE now if it
cannot deal with them properly.