Remove trailing slash: check for match
A 301 seems to imply that the redirect target exists.  Redirecting to a 404 page is all the same to a normal end user, but it seems Google will see it as a crawl error.  This change just makes sure that there is a route that will match the redirected URL; otherwise a 404 is thrown.