Avoid confusion
To me it was a bit confusing why the locale was set on the session and not the request. Further more, reading the linked article shows extra code is needed to make this work at all.

Since the linked article explains in more detail how to make the locale sticky, I suggest just setting it on the request in this code sample.