Added info on doctrine metadata implications on validation
It took me a while to figure this out. So I think it should be explained *somewhere*. I'm not sure if this is the right place, and I'm also not sure if the information is comprehensive.
Just wanted to get somehing going about 'Doctrine and validation' somewhere in the docs.