Update doctrine.rst
Note for «see doctrine documentation» just in section where `inversedBy` and `mappedBy` were just read.

I know that five sections below there is a **More Information on Associations** section where this is a reference to this (if you dig enough, because it's referenced to other page there), but that's not enough. It's too general and too late.

When you're reading this (my case, as a newcomer right now) and get to this `inversedBy` and `mappedBy` here and there isn't any direct mention about it in **Relationship Mapping Metadata** section, you felt lost. Can't understand why that is really used and what it means. And you scroll and see no direct reference about that. It breaks your learning rhythm. I had to search about this, till I found the Doctrine's documentation that I linked in that note. That cleared out the doubt. So I can continue learning. Five sections away there's a general «Read more about associations». That's ok. I will read more then. But at that moment, you need, at least, to understand, at least superficially, why that parameters are there.