Article about logout.
In this lesson, we go through the authorization HTTP BASIC. I have tried to use this example and in the end got the exact same error again and again. The Internet has a lot of open topics about this. Like many other people, I have scroll lesson up to understand where I made a mistake. I think you need to specify that this is not a working example (in this context) it will save a lot of time people who study it.