Fix Authenticator Class (getCredentials) example
The current wording surrounding the getCredentials is misleading.  Returning null does stop authentication, but it causes authentication to be successful.  The example implies that X-AUTH-TOKEN is required and should be correct.  For this behavior, either an AuthenticationException should be thrown or the credential variables should be initialized as empty and passed on to getUser().