Update formats.rst
In this case the $request->getRequestFormat() will give back "html" as default value. Possible solution is set _format as parameter in the route to overwrite the format. In my example without _format parameter it will render the article/index.html.twig and with pdf _format it will render the article/index.pdf.twig.