[PHPUnitBridge] Fix versionadded notes
I was reading the docs and noted that those features were marked as present only from Symfony 3.1/2. The reality is that the version that includes that stuff is the bridge's, which is backward compatible!
So, as an example, I can use the bridge v3.2 on my Symfony 2.8 app.

I think it should be helpfull to specify that more clearly.