Pin Python dependencies and install them in a virtualenv
Installing `sphinx` started to fail about a week ago,
with `pip` incorrectly trying to remove a Python package
shipped at the system level in Platform.sh environment.

It looks like it started failing because of a bug in
the older version of `pip` that ships with Debian Jessie.
We (Platform.sh) install global Python dependencies as
user-site packages (using `pip --user`). We also ship
system-site packages that we use for our own benefit.

So two issues here:

 * First, that older version of `pip`. We are going to
   look into upgrading it (a new batch of images based
   on Debian Stretch are cooking);
 * The build dependencies of the Symfony Docs are not
   pinned, which makes the build non-repeatable.
   I'm guessing that in this case it broke [when jinja2
   ended up requiring markupsafe>=0.23][jinja2-commit]
   (but that's just a guess).

Fix the issue for good by installing in a `virtualenv`
and pinning all the dependencies via `pip freeze`.

[jinja2-commit]: https://github.com/pallets/jinja/commit/5453db1711eeb33c68ed1126764ab691b81d1472