Update guard_authentication.rst
I still feel like the answer to "Can I have Multiple Authenticators?" is a bit confusing even with these adjustments, but I think it makes it more clear now at least what you have to adjust configuration wise. The original example simply shows the configuration for a single guard which is rather confusing considering the question it attempts to answer.

BTW; can someone else add or give me the XML change necessary? I never use XML configuration so I am not sure if it is supposed to be an attribute or another element or what not.