Update session_configuration.rst
Required changes to garbage collection documentation as it was out of date. Note that this is my first time editing and I had a look at the instructions, but some things may not be quite correct as the preview didn't seem to work properly. I didn't download the files and edit, I used the online editor. I also wasn't sure about links to other parts of the doc, but may be worth having a link to http://symfony.com/doc/current/reference/configuration/framework.html#gc-probability where I edited.