Made 'create-user' command consistent.
There is a very minor inconsistency between the naming of the 'create users' command at the beginning of the document and the 'create user' command for the rest of the article. It starts with the plural, then changes to singular. A very minor thing, but when you're following the documentation precisely, this inconsistency raises unnecessary questions. Updated all references to 'create-user'.