Resolving $this usage outside a class on form.rst
This commit take care of the framework use code examples only.

The "Creating a simple Form" define the coding place to be into a controller action method. Unlikely, the documentation consider into the other examples that the reader will know the place of the sample code. This does not allow the examples to be standalone and can lead some disappointment for which one want to have informations about a single section.

This commit offer to place the code samples into the default controller to create a logic use of the "$this" special variable.