Renamed code block 'terminal' to 'shell'
The name terminal is okay, but the generated class for the child div will be "highlight-terminal" which has a different styling than you would expect. The highlight-terminal class is used to display a shell with bullets. 

By renaming the 'terminal' to 'shell' the bullets do not show up when this code block is selected.