Added the command to install babel-preset for babel-loader
While testing this bundle i was getting hit by an erorr that i can't find the babel preset es2017 which i solved it by installing.
I'm new to this webpack thing and to work like this with javascript. Some people might be in my position and try for hours to solve this. :D