Remove promotion for the Service Locator pattern
The Service Locator is an anti-pattern that hides the true dependencies of a class, and is therefore in violation of the Law of Demeter. While it can be argued that the container should not even be in the controller in the first place, the documentation definitely should not promote SL as a way to use it, since it leads to messier and less readable code.