Post Load Event Listener
I've noticed a scenario (a bit of an edge case when uploading imagery).

If you have an entity (say a brochure) but you can give it a title and an image and the image title can be updated without the image being updated (the file type is not an UploadedFile but a File on save it'll persist the full file path instead of the desired filename. The above is a safety measure to ensure the above.