do never fetch private services from the container
Explaining that private services can still be fetched from the container
under certain circumstances does not look like a good idea. Not only is
it hard to grasp for beginners to understand why this sometimes works.
It also becomes a real problem when upgrading to Symfony 4 where all
private services won't be retrievable from the container through service
location.