Update templating.rst
I was working through the documentation at https://symfony.com/doc/current/controller.html following the Getting Started guides and that linked to this page. I considered adding this to that page but figured more people would get here from multiple places so thought this page made more sense. The instructions on how to get $this->render() to actually be available were hard to locate. I eventually found it in an explanation of how Flex aliases worked. Some of this update is taken from that example.