Fix references to annotation loader tip
Tips were refactored to an include file in
0ddf3d21b12c05fb5b34f982feb4005dab877185 but referenced the wrong
filename. a0b19c3e372d75b48e2f3606d8c4109788e9e39a removed the wrong tip
from the serializer page.

This fixes the tip references on routing/validator and restores the tip
on the serializer, but with the right reference.