Clarification on search user
I had issues with configuration because I misunderstood the configuration for the user provider. I didn't realize that the configured user was actually a static one only used for retrieving information. 

I also added a clarification for the case that you are getting information anonymously, which was what my situation required.