Update flex.rst - Add step in migration-process
I added the step to add 'symfony/symfony' to the `conflict`-section of your composer.json-file.

I did not do this until after the step to change your index.php-section.  This resulted in a lot of recipe's not being executed since the libraries for that were still installed by `symfony/symfony`.