reverted the edit on scale.rst; removed inclusion in integer.rst
As pointed out in the comments of symfony/symfony-docs/pull/9549, the scale.rst.inc file is included twice: once in integer.rst (the goal of the PR) and once in number.rst.
number.rst should not see this edit as its scale parameter is not ignored.
Hence, the inclusion was removed from integer.rst and the edit was made.
number.rst is now identical to base ref.