To help locking a command easily
There is no explained simple way of adding a lock to a command. After following this page's tutorial, the lock->acquire would always acquire and never lock even consecutive locks would not lock anything. After research, I stumbled on a suggestion on StackOverflow. It worked like a charm and it's 3 lines of code (including the "use" statement) so I think it should be in this documentation.