Update upload_file.rst
I have proposed this change: Update upload_file.rst #10143, but then realized the same wrong assumption was also part of the Entity and the comments. The wrong assumption being that the file contents are temporarily stored in the instance of the class. That's not the case. Only the full path to the temporary directory is initially stored in the instance. One can argue that that is also unnecessary, but it is a moot point as the code sets the final file name to that property once that is known.