Update parent_services.rst
The code did not correspond to the description on line 143 (of the edited version). If the setter setLogger is declared on the parent service, it will be called for all child services just as if it was in constructor and therefore lacks sense. The call should be declared on those child services, that actually use the Logger, in this case AppBundle\Repository\DoctrineUserRepository. 
This change is also in harmony with setter injection intended for optional injections as stated in https://symfony.com/doc/current/service_container/injection_types.html#setter-injection