Update forms.rst
Integrating the findings of https://github.com/EasyCorp/EasyAdminBundle/issues/2673#issuecomment-476299467

Reason for moving the block "HTML5 Validation" down:
This is just not true!: "Generated forms take full advantage of this new feature by adding sensible HTML attributes that trigger the validation." This only works if the type guesser is active, and this was only explained way later.

Reason for moving blocks "The required Option" and "label" down:
`required` was mentioned in 3 (!) places before, now it's just one.

Reason for joining "Field Type Guessing" and "Field Type Options Guessing":
The latter mainly repeated stuff already explained in the former.

Symfony's "Field Type Guessing" feature is (a) powerful and (b) obscure. It's a requirement to understand, in order to be able to understand what's going on in the form's view. Since it's now higher up on the page, more people will read it.