Creates a migration guide.
The general idea is to provide advice on how to migrate
an existing application over to Symfony using some
proven approaches.

The page should provide around 3 alternatives and some
discussion when they might be suitable as well as
additional resources like videos on this topics from
the official Symfony YouTube-channel.