[Console] Update table.rst
The Symfony\Component\Console\Helper\TableStyle class doesn't have the setDefaultCrossingChars method, but there is the setCrossingChars method. The setCrossingChars method takes at least 9 parameters, so I think it would be difficult to present it in the example instead of the setDefaultCrossingChars method. In this commit, I removed all references to the method that doesn't exist.