Update setup.rst
In my experience of the tutorial, there was no option but to configure the database URL in .env file as without doing so resulted in errors (not likely to be understood by newbies using the tutorial to learn Symfony). The need to create a database should be emphasised more here.