Inserted a warning about a limitation model transformers
There is a surprising limitation of model transformers, that cannot filter out items for a Collection field type. Took me a whole day to find out, so here is a warning about that. See https://github.com/symfony/symfony/issues/36469 for details. (Maybe that issue should be linked for further details?).