Make changes to "auto" algorithm more explicit
The only place I could find the change in behavior of `auto` for hashing algorithm (from default sodium if available to default bcrypt) was a comment within the YAML documentation example. This change in behavior should be made more explicit. (E.g., when browsing the docs looking at PHP or XML examples, this is very easy to miss).