Shortening a bit
Question: The text under the code block ("If you are using...") refers to `composer install`, not to `yarn/npm install`. So it would be cleaner to move the latter into a separate code block. But what does it actually do? In my case, `npm install` did nothing :-)

Second question: Since this is Symfony 4.4 docs, is anybody still using Symfony *without* Flex?