Update configuration.rst to aviod misunderstanding
We shouldn't deploy the ".env.local" file for the production environment variables. We just need to add the ".env" file into Git and manually create a new ".env.local" file on the production machine. In the old expression says, "deploying" and it might be a misunderstanding in this way. I changed it to "creating".

Old expression:
In production, the ``.env`` files are also parsed and loaded on each request. So
the easiest way to define env vars is by deploying a ``.env.local`` file to your
production server(s) with your production values.