Update http_kernel_controller_resolver.rst
I stumbled upon an exception:  Uncaught TypeError: LeapYearController::index(): Argument #1 ($year) must be of type int, array given, called in /var/www/web/front.php on line 32 and defined in /var/www/src/routes.php:10

This makes sense: getArguments() return indeed an array with the parameters. The index method in the LeapYearController does not expect an array at all. (new LeapYearController)->index(['year' => 2000]) is wrong, and this is happening here when an array is passed into call_user_func(). 

Fix: use spread operator for array destructuring / unpacking to pass in array values as variables. This is on PHP8.1 btw.