minor #16506 [Form] Update create_form_type_extension.rst (issamkhadiri1989)
This PR was submitted for the 6.0 branch but it was merged into the 5.4 branch instead.

Discussion
----------

[Form] Update create_form_type_extension.rst

Hello,
I think the condition `image_url is not null` is not enough as if we create a FileType form field without the `image_property` it causes an error. Adding `image_url is defined` is necessary  because the variable `image_url` is only defined and passed to the view when `image_property` is already defined.

<!--

If your pull request fixes a BUG, use the oldest maintained branch that contains
the bug (see https://symfony.com/releases for the list of maintained branches).

If your pull request documents a NEW FEATURE, use the same Symfony branch where
the feature was introduced (and `6.x` for features of unreleased versions).

-->

Commits
-------

85470c59f Update create_form_type_extension.rst