[Encore] Fox webpack-dev-server documentation
This addresses an error you get with `webpack-dev-server` where the `pfx` option actually needs a file stream to be able to work (and not just the file path). With the current example, it throws an error. The addition of the `http2` option is so that we stick better to `webpack`'s documentation (https://webpack.js.org/configuration/dev-server/#devserverhttp2) where config samples often pair `http2=true` with other `https` stuff.ñ